src/requirements.txt
Django==3.2.3
requests==2.25.1
validate_email==1.3
djangorestframework==3.12.4
python-dotenv==0.17.1
